sockets/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
pyzmq==19.0.2
requests==2.24.0